[sveltekit] Extract `useSvelteKitMutation` hook
