+++ packages/sveltekit-integration/src/lib/client/index.ts
@@ -1,0 +1,2 @@
+export * from './use-mutation.svelte';
+export * from './use-form.svelte';
